P3: Creat file p3_import_gene_info.m.
Read from 300 files in /gene_info singly and save to cell
gene_info[60*300]. Each column corresponds to the set of sites
of a gene. 60 is the maximum number of sites of a gene.

TODO: this wastes space(bacause there are many empty elements in
the cell). Maybe I should use 300 individual cells?